requirements.txt
orjson
uvloop
websockets
python-dotenv
redis==4.2.2
cassandra-driver
itsdangerous